Class 124: Insert Product. and Class 125: Error handling

+++ src/products/entities/product.entity.ts
@@ -11,7 +11,7 @@ export class Product {
   title: string;
 
   @Column({
-    type: 'numeric',
+    type: 'float',
     default: 0,
   })
   price: number;
